A working string parser that works with the import system
It can read in the user input
It changes the user input into a string vector that holds each word.
The words are then scanned one by one for key action words/ verbs like look looking use attack do open.

The parse can also read muilt word name. It can read in the names of all the doors and rooms in the world